자유게시판 - 드래그앤드롭 첨부파일 - 게시글 수정  1. View.jsp     : 사용자가 첨부파일 X버튼을 클릭하면 -> 디자인만 삭제, 삭제목록에 저장     : 사용자가 드래그앤드롭으로 새로운 첨부파일 올리면 -> Local 저장    2. 수정버튼 클릭     : Ajax로 삭제목록에 있는 첨부파일들 Local에서 삭제    3. 서버      : DB에서 게시글 수정 + filecnt도 같이 수정      : 무조건 해당 게시글의 첨부파일 목록 DB에서 삭제      : 수정버튼 클릭시 서버단으로 전송된 첨부파일이 없으면 여기서 종료      : 수정버튼 클릭시 서버단으로 전송된 첨부파일이 있으면 tbl_attach에 새로 등록

+++ jogging/src/main/webapp/WEB-INF/views/board/register.jsp
@@ -322,13 +322,13 @@ div {
 <script src="${path}/resources/js/fileAttach.js"></script>
 <script type="text/javascript">
 	
-	
+var deleteFileList = new Array();
 	var flag = '${flag}';
 	console.log('flag:' + flag);
 	
 	// Handlebars 파일템플릿 컴파일
 	var fileTemplate = Handlebars.compile($("#fileTemplate").html());
-	
+
 	
 	$(function(){		
 		// register ==> 게시글 등록과 게시글 수정
@@ -342,6 +342,9 @@ div {
 			$('.agree_btn').text('수정').css('background-color', '#FFFFFF');
 			$('.agree_btn').css('border', '1px solid #e8e8e8');
 			$('.agree_btn').css('color', '#333');
+			
+			$('.drop_txt').css('display', 'none');
+			listAttach('${path}', '${one.bno}');
 		} else if (flag == 'answer') {
 			// 답글 체크박스 죽이기
 			$('.select_box').val('${one.type}')
@@ -422,7 +425,13 @@ div {
 				});
 			
 			} else { // 게시글 수정
-				
+				var arr_size = deleteFileList.length;
+				deleteFileList[arr_size] = $(this).attr('data-src');
+				$(this).parents('li').next('input').remove();
+				$(this).parents('li').remove();
+				for(var i =0; i<deleteFileList.length;i++){
+					console.log(deleteFileList[i]);
+				}
 			}
 		});
 	});
@@ -470,11 +479,9 @@ div {
 			
 			// 로컬드라이브에 저장되어있는 해당 게시글
 			// 첨부파일 삭제
-			/* if(deleteFileList.length > 0) {
+			if(deleteFileList.length > 0) {
 				$.post('${path}/upload/deleteAllFile', {files:deleteFileList}, function(){});
-			}
-			*/
-			
+			}						
 			// 폼에 hidden 태그들을 붙임
 			$("#frm_board").append(str);
